Consolidate manager interface into 3_0
