--- src/dbus_api/api/manager_2_0/mod.rs
@@ -1,7 +1,0 @@
-mod api;
-mod methods;
-mod props;
-
-pub use api::{
-    configure_simulator_method, create_pool_method, destroy_pool_method, version_property,
-};
