--- src/dbus_api/api/manager_2_1/mod.rs
@@ -1,4 +1,0 @@
-mod api;
-mod methods;
-
-pub use api::{create_pool_method, set_key_method, unlock_pool_method, unset_key_method};
