--- src/dbus_api/api/manager_2_2/mod.rs
@@ -1,4 +1,0 @@
-mod api;
-mod methods;
-
-pub use api::set_key_method;
