--- src/dbus_api/api/manager_2_3/mod.rs
@@ -1,4 +1,0 @@
-mod api;
-mod methods;
-
-pub use api::unlock_pool_method;
