--- src/dbus_api/api/manager_2_4/mod.rs
@@ -1,4 +1,0 @@
-mod api;
-mod methods;
-
-pub use api::{create_pool_method, engine_state_report_method};
