+++ src/dbus_api/api/manager_3_0/mod.rs
@@ -1,0 +1,9 @@
+mod api;
+mod methods;
+mod props;
+
+pub use api::{
+    configure_simulator_method, create_pool_method, destroy_pool_method,
+    engine_state_report_method, set_key_method, unlock_pool_method, unset_key_method,
+    version_property,
+};
